+++ scripts/progressed/turrets/pixel/pixel-i.js
@@ -1,5 +1,5 @@
 const bitLib = this.global.pm.bitTurretLib;
-const bitBLib = require("libs/bitBulletType");
+const bitBLib = require("libs/bulletTypes/bitBulletType");
 
 const bit = bitLib.new8BitTurret("pixel-i", 8, PowerTurret, PowerTurret.PowerTurretBuild, false);
